fix: NMS class offset rounded box edges to half a pixel
batched_nms separates classes by adding class_id * 100_000 to every box
coordinate and then running one global NMS. In float32 that offset reaches
~7.9e6 at class 79, where the gap between representable values is 0.5 -- so
from class 50 upward, box edges were quantised to the nearest half pixel
before the IoU comparison. An edge at 180.25 came back as 180.0, which is
enough to flip a suppression decision sitting near the 0.45 threshold.

Doing the offset arithmetic in float64 costs nothing here: confidence
thresholding has already cut the array to tens of boxes by that point.

Re-measured with python -m src.evaluate on the 300-image val2017 subset.
The per-class table at conf 0.30 is byte-identical -- no suppression
decision flipped at the operating point -- and the mAP metrics, which gather
at conf 0.001 and so sweep the low-confidence tail, moved by 0.0001:

  mAP@0.5:0.95  0.3316 -> 0.3315
  AR@100        0.4378 -> 0.4377

Updated in results/accuracy_int8.md, README.md, and PERFORMANCE_ANALYSIS.md
sections 5.1 and 5.3. results/accuracy_int8.csv holds only the per-class
table and needed no change.

The calibration-bias table in section 5 is left as measured, with a note
saying why: both its rows predate this fix, and the
calibrated-on-the-eval-set model was a diagnostic build that is not kept, so
re-measuring one row would break a valid paired comparison.

This is not an accuracy win -- it is 0.0001 in the unfavourable direction.
The reason to keep it is that NMS now compares the coordinates it was given
rather than coordinates silently rounded for half the class list.

+++ src/detector.py
@@ -122,8 +122,13 @@ def batched_nms(
     if boxes.size == 0:
         return np.empty((0,), dtype=np.int64)
     # Larger than any plausible image dimension, so classes cannot collide.
-    offsets = class_ids.astype(np.float32) * 100_000.0
-    return nms(boxes + offsets[:, None], scores, iou_threshold)
+    # float64: at class 79 the offset is ~7.9e6, where float32 spacing is 0.5 px,
+    # so box edges would be rounded to the nearest half pixel before the IoU
+    # comparison -- enough to flip a suppression decision sitting near the
+    # threshold. There are only tens of boxes left by here, so the wider dtype
+    # is free.
+    offsets = class_ids.astype(np.float64) * 100_000.0
+    return nms(boxes.astype(np.float64) + offsets[:, None], scores, iou_threshold)
 
 
 def _make_grids(input_h: int, input_w: int) -> tuple[np.ndarray, np.ndarray]:
